chore: refaktorerer design tokens build
